pull from master and merge rotate.py into auto_augment.py
